+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Changed `multiple-columns` CSS component so its columns include a padding
 - Changed `CKEditorPlugin` for new field `variant` to choose a variant style;
+- Customize withdrawal waiver checkbox messages according to the product type
 
 ## [2.30.0] - 2024-10-16